requirements.txt
openai-harmony
python-dotenv
requests
pydantic
